Relocate i18n JS test support introduced for onboarding follows
